# docs/openapi-enterprise-v1.yaml
openapi: 3.0.3
info:
  title: WEPA Enterprise API
  version: 1.0.0
  description: Endpoints enterprise para documental, pagos, RBAC y auditoria.
servers:
  - url: http://localhost:8080/api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MobileUserRequest:
      type: object
      properties:
        username:
          type: string
          description: Si se omite, se genera desde el documento o id del integrante.
        password:
          type: string
          description: Si se omite, se genera una contrasena temporal.
        email:
          type: string
        telefono:
          type: string
    MobileUserStatusRequest:
      type: object
      required: [activo]
      properties:
        activo:
          type: boolean
    PasswordResetRequest:
      type: object
      properties:
        password:
          type: string
          description: Si se omite, se genera una contrasena temporal.
...
paths:
  /admin/integrantes/{id}/usuario:
    get:
      summary: Consultar el usuario movil asociado a un integrante.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Usuario movil asociado o estado sin usuario.
    post:
      summary: Crear usuario movil INTEGRANTE y asociarlo explicitamente al integrante.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileUserRequest"
      responses:
        "201":
          description: Usuario movil creado y credenciales temporales devueltas.
    delete:
      summary: Desasociar logicamente el usuario movil de un integrante.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Asociacion removida.
  /admin/integrantes/{id}/usuario/estado:
    patch:
      summary: Activar o desactivar el usuario movil de un integrante.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileUserStatusRequest"
      responses:
        "200":
          description: Estado del usuario actualizado.
  /admin/integrantes/{id}/usuario/password:
    patch:
      summary: Restablecer la contrasena del usuario movil de un integrante.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "200":
          description: Contrasena temporal devuelta para entrega segura al integrante.
  /admin/integrantes/{id}/foto:
    get:
      summary: Obtener foto autenticada del integrante.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Imagen del integrante.
    patch:
      summary: Actualizar foto del integrante.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [foto]
              properties:
                foto:
                  type: string
                  format: binary
      responses:
        "200":
          description: Foto actualizada.
...